Fix merge conflicts and change `earliest_year` to 1900
